Initialize file restructuring
