+++ packages/react-sdk/index.ts
@@ -1,7 +1,7 @@
 export * from '@stream-io/video-client';
 export * from '@stream-io/video-react-bindings';
 
+export * from './src/core';
+
 export * from './src/components';
-export * from './src/contexts';
-export * from './src/hooks';
 export * from './src/types';
