+++ packages/react-sdk/src/components/CallControls/ToggleAudioButton.tsx
@@ -5,7 +5,8 @@ import {
   useHasPermissions,
   useLocalParticipant,
 } from '@stream-io/video-react-bindings';
-import { useMediaDevices } from '../../contexts';
+
+import { useMediaDevices } from '../../core/contexts';
 import { DeviceSelectorAudioInput } from '../DeviceSettings';
 import { CompositeButton, IconButton } from '../Button';
 import { PermissionNotification } from '../Notification';
