+++ packages/react-sdk/src/components/DeviceSettings/DeviceSelectorVideo.tsx
@@ -1,5 +1,5 @@
 import { DeviceSelector } from './DeviceSelector';
-import { useMediaDevices } from '../../contexts';
+import { useMediaDevices } from '../../core/contexts';
 
 export type DeviceSelectorVideoProps = {
   title?: string;
