+++ packages/react-sdk/src/components/StreamCall/index.ts
@@ -1,7 +1,5 @@
-export * from './Audio';
 export * from './Stage';
 export * from './StreamCall';
-export * from './ParticipantBox';
 
 export * from './CallParticipantsView';
 export * from './CallParticipantsScreenView';
