+++ packages/react-sdk/src/components/Video/index.ts
@@ -1,4 +1,4 @@
-export * from './Video';
-export * from './BaseVideo';
-export * from './VideoPlaceholder';
+export * from '../../core/components/Video/Video';
+export * from '../../core/components/Video/BaseVideo';
+export * from '../../core/components/Video/VideoPlaceholder';
 export * from './VideoPreview';
