+++ packages/react-sdk/src/components/index.ts
@@ -1,7 +1,7 @@
 export * from './Avatar';
 export * from './Button';
 export * from './CallControls';
-export * from './CallLayout';
+export * from '../core/components/CallLayout';
 export * from './CallParticipantsList';
 export * from './CallRecordingList';
 export * from './DeviceSettings';
