+++ packages/react-sdk/src/core/components/Audio/index.ts
@@ -1,0 +1,1 @@
+export * from './Audio';
