+++ packages/react-sdk/src/core/components/CallLayout/SpeakerLayout.tsx
@@ -13,9 +13,9 @@ import {
   useParticipants,
 } from '@stream-io/video-react-bindings';
 
-import { ParticipantBox } from '../StreamCall';
-import { IconButton } from '../Button';
-import { useHorizontalScrollPosition } from '../StreamCall/hooks';
+import { ParticipantBox } from '../ParticipantBox';
+import { IconButton } from '../../../components/Button';
+import { useHorizontalScrollPosition } from '../../../components/StreamCall/hooks';
 
 export const SpeakerLayout = () => {
   const call = useCall()!;
