+++ packages/react-sdk/src/core/components/ParticipantBox/index.ts
@@ -1,0 +1,1 @@
+export * from './ParticipantBox';
