+++ packages/react-sdk/src/core/components/Video/VideoPlaceholder.tsx
