+++ packages/react-sdk/src/core/components/Video/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseVideo';
+export * from './Video';
