+++ packages/react-sdk/src/core/components/index.ts
@@ -1,0 +1,7 @@
+export * from './Audio';
+export * from './ParticipantBox';
+
+export { Video } from './Video';
+export type { VideoProps } from './Video';
+
+export * from './CallLayout';
