+++ packages/react-sdk/src/core/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useAudioPublisher';
+export * from './useVideoPublisher';
