+++ packages/react-sdk/src/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './useAudioPublisher';
-export * from './useVideoPublisher';
+export * from './useRtcStats';
